requirements/base.txt
# Core dependencies for elrond
pandas>=1.3.0,<3.0.0
openpyxl>=3.0.0,<4.0.0
pyyaml>=5.4,<7.0
python-dateutil>=2.8.0
tabulate>=0.8.9
